Move Profiler & Stopwatch into the clp namespace.
