+++ components/core/src/clp/Stopwatch.cpp
@@ -1,5 +1,6 @@
 #include "Stopwatch.hpp"
 
+namespace clp {
 Stopwatch::Stopwatch() {
     reset();
 }
@@ -23,3 +24,4 @@ double Stopwatch::get_time_taken_in_seconds() {
     std::chrono::duration<double> time_taken_in_seconds = m_time_taken;
     return time_taken_in_seconds.count();
 }
+}  // namespace clp
