+++ components/core/src/clp/clg/clg.cpp
@@ -7,11 +7,11 @@
 #include <spdlog/sinks/stdout_sinks.h>
 
 #include "../../Defs.h"
-#include "../../Profiler.hpp"
 #include "../../spdlog_with_specializations.hpp"
 #include "../GlobalMySQLMetadataDB.hpp"
 #include "../GlobalSQLiteMetadataDB.hpp"
 #include "../Grep.hpp"
+#include "../Profiler.hpp"
 #include "../streaming_archive/Constants.hpp"
 #include "../Utils.hpp"
 #include "CommandLineArguments.hpp"
@@ -21,6 +21,7 @@ using clp::GlobalMetadataDB;
 using clp::GlobalMetadataDBConfig;
 using clp::Grep;
 using clp::load_lexer_from_file;
+using clp::Profiler;
 using clp::Query;
 using clp::streaming_archive::MetadataDB;
 using clp::streaming_archive::reader::Archive;
